Refactor Test Classes
