+++ fstools/Shared/MainLayout.razor
@@ -18,17 +18,19 @@
             <div class="col-11">
                 <StopwatchComponent timerText="Timer 1" timerNumber=1 timersIn="@Timers.Timers" infoIn="@Timers.Info" fontSize="20px" />
             </div>
-@*            <div class="col-1">
-                <a href="https://docs.microsoft.com/aspnet/" target="_blank">About</a>
-            </div>*@
+ @*            <div class="col-1">
+                 <a href="https://docs.microsoft.com/aspnet/" target="_blank">About</a>
+             </div>*@
         </div>
-@*        <div class="row">
+
+         @* Show Browser Info for test *@
+@*         <div class="row">
             <div class="col-12">
                 <p>@browser.Full</p>
                 
             </div>
-        </div>*@
-
+        </div>
+ *@
         <article class="content px-4">
             @Body
         </article>
