+++ fstoolsTests/Services/BrowserServiceTests.cs
@@ -4,152 +4,151 @@ using Microsoft.JSInterop;
 using Moq;
 using Xunit;
 
-namespace fstoolsTests.Services
+namespace fstoolsTests.Services;
+
+public class BrowserServiceTests
 {
-    public class BrowserServiceTests
+    [Fact]
+    public async Task GetDataTestFull()
     {
-        [Fact]
-        public async Task GetDataTestFull()
+        const string expectedBrowserVersion = "Chrome,130.0.0.0,130,Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/130.0.0.0 Safari/537.36,";
+
+        var jsRuntimeMock = new Mock<IJSRuntime>();
+        var provider = AddServices(jsRuntimeMock);
+        var svc = provider.GetRequiredService<BrowserService>();
+
+        jsRuntimeMock
+            .Setup(j => j.InvokeAsync<string>("browserVersion", It.IsAny<object[]>()))
+            .ReturnsAsync(expectedBrowserVersion);
+        await svc.GetData();
+
+        Assert.Equal(expectedBrowserVersion, svc.Full);
+        Assert.Equal("Chrome", svc.BrowserName);
+        Assert.Equal("130.0.0.0", svc.BrowserVersion);
+        Assert.Equal("130", svc.BrowserMajorVersion);
+    }
+
+    [Fact]
+    public async Task GetDataTest2()
+    {
+        const string expectedBrowserVersion = "Chrome,130.0.0.0";
+
+        var jsRuntimeMock = new Mock<IJSRuntime>();
+        var provider = AddServices(jsRuntimeMock);
+        var svc = provider.GetRequiredService<BrowserService>();
+
+        jsRuntimeMock
+            .Setup(j => j.InvokeAsync<string>("browserVersion", It.IsAny<object[]>()))
+            .ReturnsAsync(expectedBrowserVersion);
+        await svc.GetData();
+
+        Assert.Equal(expectedBrowserVersion, svc.Full);
+        Assert.Equal("Chrome", svc.BrowserName);
+        Assert.Equal("130.0.0.0", svc.BrowserVersion);
+        Assert.Null(svc.BrowserMajorVersion);
+    }
+
+    [Fact]
+    public async Task GetDataTest1()
+    {
+        const string expectedBrowserVersion = "Chrome";
+
+        var jsRuntimeMock = new Mock<IJSRuntime>();
+        var provider = AddServices(jsRuntimeMock);
+        var svc = provider.GetRequiredService<BrowserService>();
+
+        jsRuntimeMock
+            .Setup(j => j.InvokeAsync<string>("browserVersion", It.IsAny<object[]>()))
+            .ReturnsAsync(expectedBrowserVersion);
+        await svc.GetData();
+
+        Assert.Equal(expectedBrowserVersion, svc.Full);
+        Assert.Equal("Chrome", svc.BrowserName);
+        Assert.Null(svc.BrowserVersion);
+        Assert.Null(svc.BrowserMajorVersion);
+    }
+
+    [Fact]
+    public async Task GetDataTest0()
+    {
+        const string expectedBrowserVersion = "";
+
+        var jsRuntimeMock = new Mock<IJSRuntime>();
+        var provider = AddServices(jsRuntimeMock);
+        var svc = provider.GetRequiredService<BrowserService>();
+
+        jsRuntimeMock
+            .Setup(j => j.InvokeAsync<string>("browserVersion", It.IsAny<object[]>()))
+            .ReturnsAsync(expectedBrowserVersion);
+        await svc.GetData();
+
+        Assert.Equal(expectedBrowserVersion, svc.Full);
+        Assert.Equal(string.Empty, svc.BrowserName);
+        Assert.Null(svc.BrowserVersion);
+        Assert.Null(svc.BrowserMajorVersion);
+    }
+
+    [Fact]
+    public void CanDisplayPdfTest()
+    {
+        var jsRuntimeMock = new Mock<IJSRuntime>();
+        var svc = new BrowserService(jsRuntimeMock.Object)
         {
-            const string expectedBrowserVersion = "Chrome,130.0.0.0,130,Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/130.0.0.0 Safari/537.36,";
+            BrowserName = "Chrome",
+            BrowserMajorVersion = "113"
+        };
+        Assert.False(svc.CanDisplayPdf());
 
-            var jsRuntimeMock = new Mock<IJSRuntime>();
-            var provider = AddServices(jsRuntimeMock);
-            var svc = provider.GetRequiredService<BrowserService>();
+        svc.BrowserMajorVersion = "114";
+        Assert.True(svc.CanDisplayPdf());
 
-            jsRuntimeMock
-                .Setup(j => j.InvokeAsync<string>("browserVersion", It.IsAny<object[]>()))
-                .ReturnsAsync(expectedBrowserVersion);
-            await svc.GetData();
+        svc.BrowserName = "Edg";
+        svc.BrowserMajorVersion = "113";
+        Assert.False(svc.CanDisplayPdf());
 
-            Assert.Equal(expectedBrowserVersion, svc.Full);
-            Assert.Equal("Chrome", svc.BrowserName);
-            Assert.Equal("130.0.0.0", svc.BrowserVersion);
-            Assert.Equal("130", svc.BrowserMajorVersion);
-        }
+        svc.BrowserMajorVersion = "114";
+        Assert.True(svc.CanDisplayPdf());
 
-        [Fact]
-        public async Task GetDataTest2()
-        {
-            const string expectedBrowserVersion = "Chrome,130.0.0.0";
+        svc.BrowserName = "Microsoft Internet Explorer";
+        svc.BrowserMajorVersion = "113";
+        Assert.False(svc.CanDisplayPdf());
 
-            var jsRuntimeMock = new Mock<IJSRuntime>();
-            var provider = AddServices(jsRuntimeMock);
-            var svc = provider.GetRequiredService<BrowserService>();
+        svc.BrowserMajorVersion = "114";
+        Assert.True(svc.CanDisplayPdf());
 
-            jsRuntimeMock
-                .Setup(j => j.InvokeAsync<string>("browserVersion", It.IsAny<object[]>()))
-                .ReturnsAsync(expectedBrowserVersion);
-            await svc.GetData();
+        svc.BrowserName = "Firefox";
+        svc.BrowserMajorVersion = "111";
+        Assert.False(svc.CanDisplayPdf());
 
-            Assert.Equal(expectedBrowserVersion, svc.Full);
-            Assert.Equal("Chrome", svc.BrowserName);
-            Assert.Equal("130.0.0.0", svc.BrowserVersion);
-            Assert.Null(svc.BrowserMajorVersion);
-        }
+        svc.BrowserMajorVersion = "112";
+        Assert.True(svc.CanDisplayPdf());
 
-        [Fact]
-        public async Task GetDataTest1()
-        {
-            const string expectedBrowserVersion = "Chrome";
+        svc.BrowserName = "Safari";
+        svc.BrowserMajorVersion = "14";
+        Assert.False(svc.CanDisplayPdf());
 
-            var jsRuntimeMock = new Mock<IJSRuntime>();
-            var provider = AddServices(jsRuntimeMock);
-            var svc = provider.GetRequiredService<BrowserService>();
+        svc.BrowserMajorVersion = "100";
+        Assert.False(svc.CanDisplayPdf());
 
-            jsRuntimeMock
-                .Setup(j => j.InvokeAsync<string>("browserVersion", It.IsAny<object[]>()))
-                .ReturnsAsync(expectedBrowserVersion);
-            await svc.GetData();
+        svc.BrowserMajorVersion = "99";
+        Assert.True(svc.CanDisplayPdf());
 
-            Assert.Equal(expectedBrowserVersion, svc.Full);
-            Assert.Equal("Chrome", svc.BrowserName);
-            Assert.Null(svc.BrowserVersion);
-            Assert.Null(svc.BrowserMajorVersion);
-        }
+        svc.BrowserName = "unkown";
+        Assert.False(svc.CanDisplayPdf());
 
-        [Fact]
-        public async Task GetDataTest0()
-        {
-            const string expectedBrowserVersion = "";
+        svc.BrowserName = "Chrome";
+        svc.BrowserMajorVersion = "x";
+        Assert.False(svc.CanDisplayPdf());
+    }
 
-            var jsRuntimeMock = new Mock<IJSRuntime>();
-            var provider = AddServices(jsRuntimeMock);
-            var svc = provider.GetRequiredService<BrowserService>();
+    private static ServiceProvider AddServices(Mock<IJSRuntime> mock)
+    {
+        var services = new ServiceCollection();
+        services.AddScoped<BrowserService>();
+        services.AddSingleton(mock.Object);
 
-            jsRuntimeMock
-                .Setup(j => j.InvokeAsync<string>("browserVersion", It.IsAny<object[]>()))
-                .ReturnsAsync(expectedBrowserVersion);
-            await svc.GetData();
+        var provider = services.BuildServiceProvider();
 
-            Assert.Equal(expectedBrowserVersion, svc.Full);
-            Assert.Equal(string.Empty, svc.BrowserName);
-            Assert.Null(svc.BrowserVersion);
-            Assert.Null(svc.BrowserMajorVersion);
-        }
-
-        [Fact]
-        public void CanDisplayPdfTest()
-        {
-            var jsRuntimeMock = new Mock<IJSRuntime>();
-            var svc = new BrowserService(jsRuntimeMock.Object)
-            {
-                BrowserName = "Chrome",
-                BrowserMajorVersion = "113"
-            };
-            Assert.False(svc.CanDisplayPdf());
-
-            svc.BrowserMajorVersion = "114";
-            Assert.True(svc.CanDisplayPdf());
-
-            svc.BrowserName = "Edg";
-            svc.BrowserMajorVersion = "113";
-            Assert.False(svc.CanDisplayPdf());
-
-            svc.BrowserMajorVersion = "114";
-            Assert.True(svc.CanDisplayPdf());
-
-            svc.BrowserName = "Microsoft Internet Explorer";
-            svc.BrowserMajorVersion = "113";
-            Assert.False(svc.CanDisplayPdf());
-
-            svc.BrowserMajorVersion = "114";
-            Assert.True(svc.CanDisplayPdf());
-
-            svc.BrowserName = "Firefox";
-            svc.BrowserMajorVersion = "111";
-            Assert.False(svc.CanDisplayPdf());
-
-            svc.BrowserMajorVersion = "112";
-            Assert.True(svc.CanDisplayPdf());
-
-            svc.BrowserName = "Safari";
-            svc.BrowserMajorVersion = "14";
-            Assert.False(svc.CanDisplayPdf());
-
-            svc.BrowserMajorVersion = "100";
-            Assert.False(svc.CanDisplayPdf());
-
-            svc.BrowserMajorVersion = "99";
-            Assert.True(svc.CanDisplayPdf());
-
-            svc.BrowserName = "unkown";
-            Assert.False(svc.CanDisplayPdf());
-
-            svc.BrowserName = "Chrome";
-            svc.BrowserMajorVersion = "x";
-            Assert.False(svc.CanDisplayPdf());
-        }
-
-        private static ServiceProvider AddServices(Mock<IJSRuntime> mock)
-        {
-            var services = new ServiceCollection();
-            services.AddScoped<BrowserService>();
-            services.AddSingleton(mock.Object);
-
-            var provider = services.BuildServiceProvider();
-
-            return provider;
-        }
+        return provider;
     }
 }
