+++ fstoolsTests/Services/ChartServiceTests.cs
@@ -1,33 +1,32 @@
 ﻿using fstools.Services;
 using Xunit;
 
-namespace fstoolsTests.Services
+namespace fstoolsTests.Services;
+
+public class ChartServiceTests
 {
-    public class ChartServiceTests
+    [Fact]
+    public void ChartSvcValues()
     {
-        [Fact]
-        public void ChartSvcValues()
-        {
-            var svc = new ChartService();
+        var svc = new ChartService();
 
-            Assert.Null(svc.CurrentIcao);
-            Assert.Null(svc.Pdf);
-            Assert.False(svc.ShowPdf);
+        Assert.Null(svc.CurrentIcao);
+        Assert.Null(svc.Pdf);
+        Assert.False(svc.ShowPdf);
 
-            svc.CurrentIcao = "KFFA";
-            Assert.Equal("KFFA", svc.CurrentIcao);
-            Assert.Null(svc.Pdf);
-            Assert.False(svc.ShowPdf);
+        svc.CurrentIcao = "KFFA";
+        Assert.Equal("KFFA", svc.CurrentIcao);
+        Assert.Null(svc.Pdf);
+        Assert.False(svc.ShowPdf);
 
-            svc.Pdf = "pdf path";
-            Assert.Equal("KFFA", svc.CurrentIcao);
-            Assert.Equal("pdf path", svc.Pdf);
-            Assert.False(svc.ShowPdf);
+        svc.Pdf = "pdf path";
+        Assert.Equal("KFFA", svc.CurrentIcao);
+        Assert.Equal("pdf path", svc.Pdf);
+        Assert.False(svc.ShowPdf);
 
-            svc.ShowPdf = true;
-            Assert.Equal("KFFA", svc.CurrentIcao);
-            Assert.Equal("pdf path", svc.Pdf);
-            Assert.True(svc.ShowPdf);
-        }
+        svc.ShowPdf = true;
+        Assert.Equal("KFFA", svc.CurrentIcao);
+        Assert.Equal("pdf path", svc.Pdf);
+        Assert.True(svc.ShowPdf);
     }
 }
