+++ fstoolsTests/Services/SkyVectorScrapeTests.cs
@@ -1,24 +1,23 @@
 ﻿using fstools.Services;
 using Xunit;
 
-namespace fstoolsTests.Services
-{
-    public class SkyVectorScrapeTests
-    {
-        [Fact]
-        public async Task ScrapeTest()
-        {
-            var svc = new SkyVectorScrape();
-            var charts = await svc.Scrape("KJFK");
-            Assert.NotEmpty(charts.ICAO);
-        }
+namespace fstoolsTests.Services;
 
-        [Fact]
-        public async Task ScrapeNegativeTest()
-        {
-            var svc = new SkyVectorScrape();
-            var charts = await svc.Scrape("KZZZ");
-            Assert.Empty(charts.ICAO);
-        }
+public class SkyVectorScrapeTests
+{
+    [Fact]
+    public async Task ScrapeTest()
+    {
+        var svc = new SkyVectorScrape();
+        var charts = await svc.Scrape("KJFK");
+        Assert.NotEmpty(charts.ICAO);
+    }
+
+    [Fact]
+    public async Task ScrapeNegativeTest()
+    {
+        var svc = new SkyVectorScrape();
+        var charts = await svc.Scrape("KZZZ");
+        Assert.Empty(charts.ICAO);
     }
 }
